Address PR review feedback (#694)
- Move SectionId and settingsPanelClassName into frontend/src/lib/settingsPanel.ts so SettingsPage.tsx only exports the page component (react-refresh lint)
- Cover unbroken log tokens in the LogsTab wrapping test and assert break-words on the <pre>
